+++ frontend/package.json
@@ -9,6 +9,7 @@
   },
   "dependencies": {
     "axios": "^1.6.8",
+    "bootstrap": "^5.3.3",
     "core-js": "^3.8.3",
     "vue": "^3.2.13",
     "vue-router": "^4.3.0"
