+++ frontend/src/main.js
@@ -1,6 +1,8 @@
 import { createApp } from 'vue';
 import App from './App.vue';
 import router from './router';
+import 'bootstrap/dist/css/bootstrap.min.css';
+
 
 const app = createApp(App);
 app.use(router);
